requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
bs4==0.0.2
CacheControl==0.14.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django==5.0
django-filter==24.2
django-rest-authtoken==2.1.4
django-smart-selects==1.6.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
firebase-admin==6.5.0
Flask==3.0.3
flask-marshmallow==1.2.1
flower==2.0.1
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==21.2.0
httplib2==0.22.0
humanize==4.9.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
Markdown==3.6
MarkupSafe==2.1.5
marshmallow==3.21.3
msgpack==1.0.8
packaging==24.0
peewee==3.17.5
pika==0.13.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
PyAMQP==0.1.0.7
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
str2bool==1.1
tomli==2.0.1
tornado==6.4.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
whitenoise==6.6.0
zope.interface==6.4.post2
django-celery-beat
django_extensions